// Web/RMA-Web-Shell/projects/clientcare/src/app/lead-manager/models/lead-contact-V2.ts
export class LeadContactV2 {
  contactId: number;
  leadId: number;
  name: string;
  surname: string;
  emailAddress: string;
  contactNumber: string;
  telephoneNumber: string;
  preferredCommunicationTypeId: number;
  
  createdBy: string;
  createdDate: Date;
  modifiedBy: string;
  modifiedDate: Date;
}